Create new module called motorcycle_reinforcement and move all extra taks to here
