+++ motorcycle_reinforcement/data/product_pricelist.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <data noupdate="1">
+        <record id="product_pricelist_2500_discount" model="product.pricelist">
+            <field name="name">Nuevo Cliente 2500 Descuento</field>
+        </record>
+
+        <record id="product_price_list_item_2500_discount" model="product.pricelist.item">
+            <field name="pricelist_id" ref="product_pricelist_2500_discount"/>
+            <field name="compute_price">formula</field>
+            <field name="price_discount">0</field>
+            <field name="price_surcharge">-2500</field>
+        </record>
+    </data>
+    
+</odoo>
