+++ motorcycle_stock/views/product_template_inherit.xml
@@ -33,32 +33,4 @@
             </xpath>
         </field>
     </record>
-
-    <record model="ir.ui.view" id="motorcycle_product_template_search">
-      <field name="name">motorcycle_rproduct_template_search</field>
-      <field name="model">product.template</field>
-      <field name="inherit_id" ref="product.product_template_search_view"/>
-      <field name="arch" type="xml">
-            <xpath expr="//filter[@name='services']" position="after">
-                <filter string="Motocicletas" name="motorcycle" domain="[('detailed_type','=','motorcycle')]"/>
-            </xpath>
-      </field>
-    </record>
-
-    <record id="sale.product_template_action" model="ir.actions.act_window">
-        <field name="name">Products</field>
-        <field name="res_model">product.template</field>
-        <field name="view_mode">kanban,tree,form,activity</field>
-        <field name="view_id" ref="product.product_template_kanban_view"/>
-        <field name="search_view_id" ref="product.product_template_search_view"/>
-        <field name="context">{"search_default_motorcycle":1, "sale_multi_pricelist_product_template": 1}</field>
-        <field name="help" type="html">
-            <p class="o_view_nocontent_smiling_face">
-                Create a new product
-            </p><p>
-                You must define a product for everything you sell or purchase,
-                whether it's a storable product, a consumable or a service.
-            </p>
-        </field>
-    </record>
 </odoo>
